move all sources to src, add babel build
