+++ .gitignore
@@ -3,3 +3,4 @@ html-report
 node_modules
 tmp
 .idea
+/lib
