+++ package.json
@@ -24,32 +24,40 @@
     "url": "https://github.com/SitePen/remap-istanbul"
   },
   "license": "BSD-3-Clause",
-  "main": "main.js",
-  "bin": "./bin/remap-istanbul",
+  "main": "./lib/main.js",
+  "bin": "./lib/remap-istanbul.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/SitePen/remap-istanbul.git"
   },
   "scripts": {
-    "lint": "eslint ./lib/",
-    "test": "tests/run.sh"
+    "lint": "eslint ./src/",
+    "test": "tests/run.sh",
+    "build": "rimraf lib && babel --presets es2015 --plugins transform-runtime src --out-dir ./lib --source-maps",
+    "prepublish": "npm run build"
   },
   "dependencies": {
     "amdefine": "1.0.0",
+    "babel-runtime": "^6.11.6",
     "gulp-util": "3.0.7",
     "istanbul": "0.4.5",
     "source-map": ">=0.5.6",
     "through2": "2.0.1"
   },
   "devDependencies": {
+    "babel-cli": "^6.16.0",
+    "babel-core": "^6.17.0",
+    "babel-plugin-transform-runtime": "^6.15.0",
+    "babel-preset-es2015": "^6.16.0",
+    "codecov.io": "0.1.6",
     "eslint": "^2.2.0",
     "eslint-config-airbnb": "^9.0.1",
     "eslint-plugin-import": "^1.7.0",
     "eslint-plugin-jsx-a11y": "^1.2.0",
     "eslint-plugin-react": "^5.0.1",
-    "codecov.io": "0.1.6",
     "grunt": "^1.0.1",
     "gulp": "3.9.1",
-    "intern": "^3.3.0"
+    "intern": "^3.3.0",
+    "rimraf": "^2.5.4"
   }
 }
