+++ src/CoverageTransformer.js
