+++ src/SparceCoverageCollector.js
