+++ src/gulpRemapIstanbul.js
