+++ src/intern-reporters/JsonCoverage.js
