+++ src/loadCoverage.js
